Move isNotContainedSameMts method to helper
